refactor(memory): rename the extraction module to writer
Part of #599
