+++ src/discordbot/services/memory/writer.py
